Feature: Added ability to parse entire table, returning several course results by reading each row. -This is returned as a slice of courses.
Additional: Began adding unit tests into the project. Should consider thoroughly testing each method and error catching.

A few errors were run into, with degenerate cases where the table rows were formatted strangely.
1. TimeBegin and TimeEnd for a course were merged into a single cell, indicating --ARR--, likely meaning it was TBD. Used a specific case to solve this.
2. Courses sometimes contained Additional Times as an extra row. I remedied this by adding a new struct for additional times.
This also required that we parse ahead and check in the tokenizer buffer if it contained the term "Additional Times" in it.

These are very fragile and only tested with inputs that I am aware of that may have these degenerate cases. It is likely that with a site update to the TimeTable, this will be deprecated.